use Rc to manage data in Image<T>

+++ src/geo/mod.rs
@@ -5,7 +5,7 @@ pub mod rect;
 pub use self::point::Point;
 pub use self::rect::Rect;
 
-pub type Pointi = Point<i32>;
+pub type Pointi = Point<isize>;
 pub type Pointf = Point<f32>;
-pub type Recti = Rect<i32>;
+pub type Recti = Rect<isize>;
 pub type Rectf = Rect<f32>;
